requirements.txt

pillow
scikit-learn
numpy==1.26.4
pandas